+++ src/include/wanted-api.h
@@ -114,6 +114,7 @@ typedef struct wantedConfig_t {
 
 int WantedWappRun(wapp_data_t *ctx);
 void WantedWappStop(wapp_data_t *ctx);
+void WantedWappTerminate(wapp_data_t *ctx);
 int WantedWappParseManifest(wapp_t *w);
 int WantedWappParseManifestBytes(wapp_t *w, const uint8_t *manifest,
                                  size_t manifestLen);